# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Auth Service Swagger
  version: 1.0.0
  description: API documentation for the Auth Service

servers:
  - url: http://localhost:5001/service/auth

paths:
  /login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: encrypted_password
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Invalid input
        "500":
          description: Internal server error

  /register:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        "200":
          description: Successful registration
        "400":
          description: Invalid input
        "500":
          description: Internal server error

  /logout:
    post:
      summary: User logout
      responses:
        "200":
          description: Successful logout
        "500":
          description: Internal server error

  /reset-password:
    post:
      summary: Reset user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
      responses:
        "200":
          description: Password reset email sent
        "400":
          description: Invalid input
        "500":
          description: Internal server error

  /change-password:
    post:
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: old_password123
                new_password:
                  type: string
                  example: new_password123
      responses:
        "200":
          description: Password changed successfully
        "400":
          description: Invalid input
        "500":
          description: Internal server error

  /deactivate-account:
    post:
      summary: Deactivate user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: johndoe123
                password:
                  type: string
                  example: encrypted_password
              required:
                - username
                - password
      responses:
        "200":
          description: Account deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Account deactivated successfully
        "400":
          description: Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid username or password
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
  /health:
    get:
      summary: Health check for Auth Service
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK